requirements.txt
streamlit
numpy
seaborn
plotly.express
scipy